fix(misleading-error): allow to return better error messages
The point of this change is to return better error message in case name or direction haven't been specified into the directive.

If nothing was specified the filter was returning an empty array, and gettting the [0] element was getting undefined in variable direction, and an exception was thrown at the line `return direction.value.value;` : TypeError: Cannot read property 'value' of undefined.

The same apply to getRelationDirection and getRelationName method.