Temporal type API refactor and update
The translation process for temporal types no longer relies on the use of the Cypher toInteger function to serialize values appropriately for the temporal functions. Reducing the complexity of the translation by using neo4j-driver to prepare the .year, .time, etc. as Integers, allowed for the support of list type fields of temporal types, on both nodes and relationships, using the reduce list function.

Various complex branches have been added for the determination of the appropriate Cypher variable names throughout the translation cases, I hope to clean them up and comment them some more pretty soon