Add CONTRIBUTING.md
I'm trying to contribute a fix for https://github.com/neo4j-graphql/neo4j-graphql-js

When I am about to contribute to a software library, the first thing I do is to
look for software tests and run them. Unfortunately, there is no documentation
how to setup tests.

What I did:

* `npm install` - intuitive
* I checked `package.json` - aha, there is `npm run test` and `npm run test-all`
* When I run `npm run test-all` all the integration tests fail. Anything
  missing? Checking CircleCI https://circleci.com/gh/neo4j-graphql/neo4j-graphql-js/3519
  reveals I have to `node -r babel-register ./example/apollo-server/movies-middleware.js`
* Scrolling down reveals `npm run parse-tck` :frowning:
* And finally `./node_modules/.bin/nyc ./node_modules/.bin/ava test/*.js test/neo4j-schema/*Test.js`

Ok, so first suggestion: I would expect all the above instruction to go into
`CONTRIBUTING.md`. Alternatively it should be in `README.md` in a section
`Contributing`. I checked the other projects of neo4j here is what I discovered:

The following repos of `neo4j` use the exact same `CONTRIBUTING.md`:
* https://github.com/neo4j/neo4j/blob/3.5/CONTRIBUTING.md
* https://github.com/neo4j/neo4j-java-driver/blob/2.0/CONTRIBUTING.md
* https://github.com/neo4j/neo4j-javascript-driver/blob/1.7/CONTRIBUTING.md
* https://github.com/neo4j/neo4j-go-driver/blob/1.7/CONTRIBUTING.md

So that's why I decided to commit the same file to the repository. I find the
link to the neo4j-users slack helpful but the rest of it is pretty generic and
does not explain how to setup tests. So where does that information go?

Next up: Why are your `npm` scripts different from what you use on CircleCI? For
me the build server is the single source of truth. It confuses me that CircleCI
uses a different set of tests than your `npm run test-all` for example. Are all
the tests supposed to pass, after all? On my machine, plenty of integration
tests fail and I'm confused if they might be expected to fail.

So, dear contributors, @johnymontana and @#2michaeldgraham how do I setup tests?

Cheers,
Robert