Use `debug` for debugging
The motivation for this change is that we have a lot of the following
code:
```js
SomeResolver: async (object, args, context, resolveInfo) => {
  ...
  return neo4jgraphql(object, args, context, resolveInfo, false)
}
```

As soon as we want to debug the generated cypher queries, we only see those
of the augmented resolvers, but not if the call to `neo4jgraphql` happens
in our own custom resolver.

With this commit you can do the following:
```sh
DEBUG=neo4j npm run something
```
and see all the generated cypher queries.

Of course, this can be improved: It would be nice to selectively
filter for certain resolvers like this:
```sh
DEBUG=neo4j:CreatePost npm run something
```
and see only the generated cypher queries for the `CreatePost` resolver.

But that's for another day
- [KISS](https://en.wikipedia.org/wiki/KISS_principle) for now.