Fix for incorrect behaviour in _some filtering
FIx for  incorrect behaviour in _some filtering for filters on types below the root. Variable created from type-name is replaced with the variableName the function receives.

A variable that is derived from typename means that any path-selection will be performed using the full set of nodes of this type as left set. When filters are defined on relationships we actually want to filter on the set that is represented on the parent relationship. See issue #326