Workaround for Neo4j 4.0 database load
For historical reasons we copy the graph.db directory for the
test database instead of using neo4j-admin load. This workaround
ensures the test database will start when databases are loaded
this way. This is a temporary workaround and the test database
will be loaded via a Cypher script in the future.