Fix for unused generated mutation payload types
The augmentation process was generating output types for relationship update mutations even when that relationship has no properties to update. Some logic was previously added to prevent the generation of such an update mutation field, but the same logic was not yet applied to the generation of associated output types.