Fixes #446
Removes decideRelationTypeDirection as a result of refactoring the decision it was making when filtering the incoming direction of a relationship type. It had an undefined input argument for such cases, so a comparison was failing to ever result in swapping the variables for .from / .to nodes and reversing the relationship direction. This is done due to the use of a more static template once these variables get used to produce Cypher.

Also removes a few unused function arguments