Fix for issue #446 (#469)
* Fixes #446

Removes decideRelationTypeDirection as a result of refactoring the decision it was making when filtering the incoming direction of a relationship type. It had an undefined input argument for such cases, so a comparison was failing to ever result in swapping the variables for .from / .to nodes and reversing the relationship direction. This is done due to the use of a more static template once these variables get used to produce Cypher.

Also removes a few unused function arguments

* Adds test for #446 fix

Issue #446 occured when filtering incoming relationships through a relationship type. This case was not tested in the TCK. Only outgoing cases were tested when querying from Person to Company, filtering over the `Person.employmentHistory` field.

So a new test has been added that covers the reverse, querying from Company to Person, filtering over the `Company.employeeHistory` field.